climbing fixed
Added Climb provider to XR origin. So we can use the prefb
Final handles were added so player can move forward to a platform.
Player does not go through floor while climbing becuase prefab allowed him to do so.
The only hting that is missing is configuruing the handles so they change color on hover, select when player interacts with them.